change localstorage to usestate
