+++ pages/index.js
@@ -1,16 +1,49 @@
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import useSWR from "swr";
 import EventCard from "../components/EventCard";
 import StyledLink from "../components/StyledLink";
 import Svg from "../components/Svg";
 import { fetcher } from "../utils/api";
 
-export default function Home({ attendances, setAttendances }) {
+export default function Home() {
+  const [attendances, setAttendances] = useState([]);
   const { data: players } = useSWR("/api/players", fetcher);
   const { data: events, error } = useSWR(players ? "/api/events" : null, fetcher, {
     onSuccess: (data) => {
-      for (const event of data) {
-        setAttendances(players, event.id);
+      for (const oneEvent of data) {
+        const acceptedPlayers = players.filter((player) =>
+          player.attendances.find(
+            (attendance) =>
+              Object.values(attendance).includes(oneEvent.id) && Object.values(attendance).includes("accepted")
+          )
+        );
+        const cancelledPlayers = players.filter((player) =>
+          player.attendances.find(
+            (attendance) =>
+              Object.values(attendance).includes(oneEvent.id) && Object.values(attendance).includes("cancelled")
+          )
+        );
+        setAttendances((current) => {
+          const currentEvent = current.find((event) => event.eventId === oneEvent.id);
+          if (currentEvent) {
+            current.map((att) => {
+              if (att.eventId === currentEvent.eventId) {
+                att.accepted = acceptedPlayers.length;
+                att.cancelled = cancelledPlayers.length;
+                return att;
+              } else return att;
+            });
+            return [...current];
+          }
+          return [
+            ...current,
+            {
+              eventId: oneEvent.id,
+              accepted: acceptedPlayers.length,
+              cancelled: cancelledPlayers.length,
+            },
+          ];
+        });
       }
     },
   });
